requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dropbox==10.3.1
Flask==1.1.2
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
msgpack==1.0.0
neovim==0.3.1
parso==0.7.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pynvim==0.4.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
